+++ src/main/java/dk/superawesome/factorio/mechanics/impl/accessible/Constructor.java
@@ -232,7 +232,7 @@ public class Constructor extends AbstractMechanic<Constructor> implements Access
 
     @Override
     public List<ItemStack> take(int amount) {
-        return this.<ConstructorGui>take(Math.min(getMaxTransfer(), amount), storageType, storageAmount, getGuiInUse(), ConstructorGui::updateRemovedItems, new HeapToStackAccess<>() {
+        return this.<ConstructorGui>take((int) Math.min(getMaxTransfer(), amount), storageType, storageAmount, getGuiInUse(), ConstructorGui::updateRemovedItems, new HeapToStackAccess<>() {
             @Override
             public Integer get() {
                 return storageAmount;
@@ -256,12 +256,12 @@ public class Constructor extends AbstractMechanic<Constructor> implements Access
     }
 
     @Override
-    public int getMaxTransfer() {
+    public double getMaxTransfer() {
         return storageType.getMaxStackSize();
     }
 
     @Override
-    public int getTransferAmount() {
+    public double getTransferAmount() {
         return storageAmount;
     }
